Refactor initial syncer classes (#233)
